+++ src/components/hooks/useForm/FormContext.ts
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+
+import { IUseFormValues } from './useForm';
+
+export const FormContext = createContext<IUseFormValues | undefined>(undefined);
