+++ src/components/hooks/useForm/useFieldContext.ts
@@ -1,0 +1,3 @@
+import { useFormContext } from './useFormContext';
+
+export const useFieldContext = (name: string) => useFormContext().getFieldProps(name);
