+++ src/components/hooks/useForm/useFormContext.ts
@@ -1,13 +1,11 @@
 import { useContext } from 'react';
 
-import FormContext from './FormContext';
+import { FormContext } from './FormContext';
 
-const useFormContext = () => {
+export const useFormContext = () => {
   const context = useContext(FormContext);
   if (context === undefined) {
     throw new Error('useFormContext must be used within a Form');
   }
   return context;
 };
-
-export default useFormContext;
