+++ src/components/organisms/Form/Form/Form.tsx
@@ -1,6 +1,5 @@
 import React, { FC, HTMLAttributes } from 'react';
-import { FormContext } from '../context';
-import useForm, { IUseFormProps } from '../../../hooks/useForm/useForm';
+import { useForm, IUseFormProps, FormContext } from '../../../hooks/useForm';
 
 const Form: FC<Omit<HTMLAttributes<HTMLFormElement>, 'onSubmit'> & IUseFormProps> = ({
   children,
