+++ src/components/organisms/Form/FormButton/FormButton.tsx
@@ -1,7 +1,6 @@
 import React, { FC } from 'react';
 import Button, { IButtonProps } from '../../../atoms/Button/Button';
-
-import { useFormContext } from '../context';
+import { useFormContext } from '../../../hooks/useForm';
 
 export interface IFromButtonProps extends IButtonProps {
   type?: 'button' | 'reset' | 'submit';
