--- src/components/organisms/Form/context/FormContext.ts
@@ -1,7 +1,0 @@
-import { createContext } from 'react';
-
-import { IUseFormValues } from '../../../hooks/useForm/useForm';
-
-const FormContext = createContext<IUseFormValues | undefined>(undefined);
-
-export default FormContext;
