--- src/components/organisms/Form/context/useFieldContext.ts
@@ -1,5 +1,0 @@
-import useFormContext from './useFormContext';
-
-const useFieldContext = (name: string) => useFormContext().getFieldProps(name);
-
-export default useFieldContext;
